Move latest_reports_overview to external.database.
